forwarding address code cleanup
